refactor: extractQueryParameter() 메서드를 RequestLine 클래스로 이동"
